+++ lib/commons.dart
@@ -12,8 +12,10 @@ export 'package:commons/src/screens/stack_trace_dialog.dart';
 export 'package:commons/src/functions/functions.dart';
 export 'package:commons/src/functions/extension_functions.dart';
 export 'package:commons/src/functions/navigation_functions.dart';
+export 'package:commons/src/shared_preferences/shared_prefrences.dart';
 
 // Dependencies export
 export 'package:url_launcher/url_launcher.dart';
 export 'package:device_info/device_info.dart';
 export 'package:share/share.dart';
+export 'package:shared_preferences/shared_preferences.dart';